Replace `sprintf-js` with a custom handler
Sorry for making this change a while after opening the pr, but `sprintf-js` has some things that translations shouldn't support (such as `%j`), and it uses `undefined` for missing entries instead of empty strings (like translations), so it was replaced with a custom formatting handler. Also, `toString` and `toMotd` were rewritten to handle `translate` components without `with` as if they had `[]` as their with, instead of just appending the translation format to the string.