Better translation handling (#83)
* Improve how `translate` components are handled

In the `toString` and `toMotd` methods of `ChatMessage`, `lang[this.translate] ?? this.translate` is now used to get the translation value (correctly), and uses of vsprintf are now try-catch'ed to handle invalid translations (`translation.test.invalid`, `%`, etc...)

* Use a `translate` component for `fromNetwork`

* Tests

* Invalid test

i forgor :skull:

* `message += format`

* Replace `sprintf-js` with a custom handler

Sorry for making this change a while after opening the pr, but `sprintf-js` has some things that translations shouldn't support (such as `%j`), and it uses `undefined` for missing entries instead of empty strings (like translations), so it was replaced with a custom formatting handler. Also, `toString` and `toMotd` were rewritten to handle `translate` components without `with` as if they had `[]` as their with, instead of just appending the translation format to the string.